// src/types/application.types.ts
export type UpdateUpgradeApplicationPayload = {
  targetTier?: string;
  autorizedPersonFullName?: string;
  authorizedPersonTitle?: string;
  companyName?: string;
  tradingName?: string;
  Reg_no?: string;
  country?: string;
  address?: string;
  email?: string;
  phone?: string;
  website?: string;
  type?: string;
  operation_mode?: string;
  yearsInOperation?: string;
  contactName?: string;
  contactPosition?: string;
  contactPhone?: string;
  contactEmail?: string;
  businessNature?: string[];
  countriesOperateFrom?: string;
  countriesShipTo?: string;
  cargoTypes?: string[];
  estimatedMonthlyVolume?: string;
  servicesRequired?: string;
  status?: string;
  notes?: string;
};

export type UpdateHubProviderApplicationPayload = {
  shopName?: string;
  address?: string;
  landmark?: string;
  cityOrState?: string;
  contact?: string;
  email?: string;
  cctvAvailable?: boolean;
  ownerName?: string;
  ownerEmail?: string;
  prefferedContactMethod?: string;
  operatingDays?: string[];
  email_active_window_from?: string;
  email_active_window_to?: string;
  daily_minimum_staff?: number;
  daily_maximum_staff?: number;
  daily_foot_traffic?: string;
  handledDeliveryServiceBefore?: boolean;
  atLeastSixMonthCommitted?: boolean;
  comments?: string;
  image_urls?: string[];
  status?: string;
  rejection_notes?: string;
};

export type UpdateCorporateApplicationPayload = {
  companyName?: string;
  tradingName?: string;
  regNo?: string;
  country?: string;
  address?: string;
  yearsInOperation?: string;
  contactName?: string;
  contactPosition?: string;
  contactPhone?: string;
  contactEmail?: string;
  website?: string;
  businessNature?: string[];
  countriesOperateFrom?: string;
  countriesShipTo?: string;
  cargoTypes?: string[];
  estimatedMonthlyVolume?: string;
  servicesRequired?: string;
  status?: string;
  rejectionNotes?: string;
};